Moved reusable functions to AccessibilityHelpers and removed erroneous VectorField grid size
